Move hfp fetch and caching to worker thread WIP.
