+++ .babelrc
@@ -1,6 +1,7 @@
 {
   "presets": [["react-app", {"modules": false}]],
   "plugins": [
+    "@babel/plugin-transform-runtime",
     "react-hot-loader/babel",
     ["@babel/plugin-proposal-decorators", {"legacy": true}],
     ["@babel/plugin-proposal-class-properties", {"loose": true}],
