+++ package.json
@@ -4,6 +4,8 @@
   "private": true,
   "dependencies": {
     "@babel/core": "^7.2.2",
+    "@babel/plugin-transform-runtime": "^7.2.0",
+    "@babel/runtime": "^7.2.0",
     "@svgr/webpack": "2.4.1",
     "apollo-cache-inmemory": "^1.3.12",
     "apollo-client": "^2.4.8",
@@ -91,7 +93,7 @@
     "webpack-dev-server": "^3.1.14",
     "webpack-manifest-plugin": "^2.0.4",
     "workbox-webpack-plugin": "^3.6.3",
-    "worker-loader": "^2.0.0"
+    "workerize-loader": "^1.0.4"
   },
   "scripts": {
     "docker:build": "docker build -t hsldevcom/transitlog-ui .",
