+++ src/helpers/hfpQueryManager.js
@@ -7,6 +7,7 @@ import lruDriver from "localforage-lru-driver";
 import {combineDateAndTime} from "./time";
 import pAll from "p-all";
 import moment from "moment-timezone";
+import {createHfpItem} from "./createHfpItem";
 
 // Bump db version if you change something concerning the local cache.
 // That will make all client's databases clear out.
@@ -26,21 +27,6 @@ const localforage = indexedLocalforage.defineDriver(lruDriver).then(async () => 
 
   return lf;
 });
-
-// Add props to or modify the HFP item.
-export function createHfpItem(rawHfp) {
-  const journeyStartMoment = combineDateAndTime(
-    rawHfp.oday,
-    rawHfp.journey_start_time,
-    "Europe/Helsinki"
-  );
-
-  return {
-    ...rawHfp,
-    received_at_unix: moment.tz(rawHfp.received_at, "Europe/Helsinki").unix(),
-    journey_start_timestamp: journeyStartMoment.toISOString(),
-  };
-}
 
 // Loads localstorage cache into memory
 export async function loadCache() {
